pipeline/requirements.txt
psycopg2-binary==2.9.9
pytest==8.1.1
pytest-cov==5.0.0

# Kafka consumer simulation
kafka-python==2.0.2
